+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Change Log
 HEAD
 ----
 - Fixed `Protocol.get_free_busy_info()` when called with +100 accounts.
+- Allowed configuring DNS timeout for a single nameserver
+  (`Autodiscovery.DNS_RESOLVER_ATTRS["timeout""]`) and the total query lifetime
+  (`Autodiscovery.DNS_RESOLVER_LIFETIME`) separately.
 
 
 4.7.4